Criando classe "Buscador"

+++ src/main/java/br/gov/servicos/dominio/Busca.java
@@ -6,7 +6,7 @@ import org.springframework.data.elasticsearch.annotations.Document;
 import org.springframework.data.elasticsearch.annotations.Field;
 
 import static org.springframework.data.elasticsearch.annotations.FieldIndex.not_analyzed;
-import static org.springframework.data.elasticsearch.annotations.FieldType.Long;
+import static org.springframework.data.elasticsearch.annotations.FieldType.Integer;
 
 @Value
 @Document(indexName = "guia-de-servicos", type = "busca")
@@ -15,17 +15,17 @@ public class Busca {
     @Id
     String termo;
 
-    @Field(index = not_analyzed, type = Long)
-    Long resultados;
+    @Field(index = not_analyzed, type = Integer)
+    Integer resultados;
 
-    @Field(index = not_analyzed, type = Long)
-    Long ativacoes;
-
+    @Field(index = not_analyzed, type = Integer)
+    Integer ativacoes;
+    
     private Busca() {
         this(null, null, null);
     }
 
-    public Busca(String termo, Long resultados, Long ativacoes) {
+    public Busca(String termo, Integer resultados, Integer ativacoes) {
         this.termo = termo;
         this.resultados = resultados;
         this.ativacoes = ativacoes;
